Progress on actually outputing sounds

+++ src/apu/mixer.rs
@@ -1,19 +1,19 @@
 // https://wiki.nesdev.com/w/index.php/APU_Mixer#Emulation
 pub struct Mixer {
-    pulse_table: [f64; 31],
-    tnd_table: [f64; 203],
+    pulse_table: [f32; 31],
+    tnd_table: [f32; 203],
 }
 
 impl Mixer {
     pub fn new() -> Self {
         let mut pulse_table = [0.0; 31];
         for (index, val) in pulse_table.iter_mut().enumerate() {
-            *val = 95.52 / (8128.0 / index as f64 + 100.0);
+            *val = 95.52 / (8128.0 / index as f32 + 100.0);
         }
 
         let mut tnd_table = [0.0; 203];
         for (index, val) in tnd_table.iter_mut().enumerate() {
-            *val = 163.67 / (24329.0 / index as f64 + 100.0)
+            *val = 163.67 / (24329.0 / index as f32 + 100.0)
         }
 
         Mixer {
@@ -29,7 +29,7 @@ impl Mixer {
         triangle_output: u8,
         noise_output: u8,
         dmc_output: u8,
-    ) -> f64 {
+    ) -> f32 {
         let pulse_table_index = (pulse_1_output + pulse_2_output) as usize;
         let pulse_out = self.pulse_table[pulse_table_index];
         let tnd_table_index = (3 * triangle_output + 2 * noise_output + dmc_output) as usize;
